+++ src/ZfrStripe/Client/ServiceDescription/Stripe-2014-03-28.php
@@ -131,6 +131,7 @@ return array(
                     'description' => 'Whether or not to immediately capture the charge',
                     'location'    => 'query',
                     'type'        => 'boolean',
+                    'filters'     => array('ZfrStripe\Client\Filter\BooleanFilter::encodeValue'),
                     'required'    => false
                 ),
                 'description' => array(
@@ -256,6 +257,7 @@ return array(
                     'description' => 'Indicate whether the application fee should be refunded when refunding this charge',
                     'location'    => 'query',
                     'type'        => 'boolean',
+                    'filters'     => array('ZfrStripe\Client\Filter\BooleanFilter::encodeValue'),
                     'required'    => false
                 ),
                 'expand' => array(
@@ -752,6 +754,7 @@ return array(
                     'description' => 'A flag that if set to true will delay the cancellation of the subscription until the end of the current period.',
                     'location'    => 'query',
                     'type'        => 'boolean',
+                    'filters'     => array('ZfrStripe\Client\Filter\BooleanFilter::encodeValue'),
                     'required'    => false
                 ),
                 'expand' => array(
@@ -934,6 +937,7 @@ return array(
                     'description' => 'Flag telling us whether to prorate switching plans during a billing cycle',
                     'location'    => 'query',
                     'type'        => 'boolean',
+                    'filters'     => array('ZfrStripe\Client\Filter\BooleanFilter::encodeValue'),
                     'required'    => false
                 ),
                 'trial_end' => array(
@@ -1560,6 +1564,7 @@ return array(
                     'description' => 'Boolean representing whether an invoice is closed or not',
                     'location'    => 'query',
                     'type'        => 'boolean',
+                    'filters'     => array('ZfrStripe\Client\Filter\BooleanFilter::encodeValue'),
                     'required'    => false
                 ),
                 'expand' => array(
@@ -2128,6 +2133,7 @@ return array(
                     'description' => 'Boolean to only return recipients that are verified or unverified',
                     'location'    => 'query',
                     'type'        => 'boolean',
+                    'filters'     => array('ZfrStripe\Client\Filter\BooleanFilter::encodeValue'),
                     'required'    => false
                 ),
                 'expand' => array(